Configure `Local` environment
I defined new environment called "Local". This environment is set
by default in `launchsettings.json`. This file in general should not
be added to .gitignore and should be the same for all developers to
keep consistent port numbers and environment settings on each machine.

When environment is set to "Local" then ParameterStore is not added
to configuration sources at all. Otherwise ParameterStore path is set
to `/FeatureToggleDemo/{envName}` so parameter path would be for example:
`FeatureToggleDemo/Staging/FeatureToggles/IsFirstFeatureEnabled`.

When app is configured like this we should be able to run it locally
as well as deploy to any environment without need of modifications.
Only thing we need to do is set proper `ASPNET_ENVIRONMENT` variable
and override needed parameters in ParameterStore using matching
environment name in parameter path.